pyproject.toml
[project]
name = "agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "fastapi[standard]>=0.125.0",
    "openai-agents>=0.6.3",
    "python-dotenv>=1.2.1",
]
